+++ src/test/java/com/github/zeykrus/bankpet/TestConstants.java
@@ -1,10 +1,10 @@
 package com.github.zeykrus.bankpet;
 
 public class TestConstants {
-    public static final double POSITIVE_AMOUNT = 10;
-    public static final double BIG_POSITIVE_AMOUNT = 10000;
-    public static final double ZERO_AMOUNT = 0;
-    public static final double NEGATIVE_AMOUNT = -10;
-    public static final double BIG_NEGATIVE_AMOUNT = -10000;
+    public static final long POSITIVE_AMOUNT = 10;
+    public static final long BIG_POSITIVE_AMOUNT = 10000;
+    public static final long ZERO_AMOUNT = 0;
+    public static final long NEGATIVE_AMOUNT = -10;
+    public static final long BIG_NEGATIVE_AMOUNT = -10000;
     public static final String PERSON_OWNER = "Person Owner1";
 }
